% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_plot.R
\name{l2_plot}
\alias{l2_plot}
\title{Plot Function of Shadowing Improvement
This function plots improvement of shadowing experiments, by comparing the subjects' IPA transcription with an assumed "standard" IPA.}
\usage{
l2_plot(reshaped_df)
}
\arguments{
\item{reshaped_df}{a dataframe processed by the l2_reshape function.}
}
\description{
Plot Function of Shadowing Improvement
This function plots improvement of shadowing experiments, by comparing the subjects' IPA transcription with an assumed "standard" IPA.
}
\examples{
l2_plot()
}
