% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_preprocess.R
\name{l2_preprocess}
\alias{l2_preprocess}
\title{IPA Preprocessor
This function normalizes potential mistakes made by IPA transcribers.}
\usage{
l2_preprocess(df)
}
\arguments{
\item{df}{a dataframe which follows this package's template.}
}
\description{
IPA Preprocessor
This function normalizes potential mistakes made by IPA transcribers.
}
\examples{
l2_preprocess()
}
