% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_reshape.R
\name{l2_reshape}
\alias{l2_reshape}
\title{Shadowing CSV Reshaper
This function reshapes the CSV file of the shadowing experiment to a processable format.}
\usage{
l2_reshape(df)
}
\arguments{
\item{df}{a dataframe that follows the template required of this package.}
}
\description{
Shadowing CSV Reshaper
This function reshapes the CSV file of the shadowing experiment to a processable format.
}
\examples{
l2_reshape()
}
