% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relapse_detect.R
\name{relapse_detect}
\alias{relapse_detect}
\title{Relapse Detection
This function detects when an error rate goes up again mid-trial, which we take to be a sign of relapse}
\usage{
relapse_detect(subset_word)
}
\arguments{
\item{subset_word}{}
}
\value{

}
\description{
Relapse Detection
This function detects when an error rate goes up again mid-trial, which we take to be a sign of relapse
}
