% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_error.R
\name{t_error}
\alias{t_error}
\title{Pronunciation Error Rate
This function retrieves the ordered error rate in any given trial}
\usage{
t_error(reshaped_df, t_number)
}
\arguments{
\item{t_number}{}
}
\value{

}
\description{
Pronunciation Error Rate
This function retrieves the ordered error rate in any given trial
}
