% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_plot.R
\name{word_plot}
\alias{word_plot}
\title{Plot Function of Single Word Shadowing Improvement
This function plots the shadowing improvement of a single word}
\usage{
word_plot(reshaped_df, word_to_plot)
}
\arguments{
\item{word_to_plot}{}
}
\value{

}
\description{
Plot Function of Single Word Shadowing Improvement
This function plots the shadowing improvement of a single word
}
