% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_subset.R
\name{word_subset}
\alias{word_subset}
\title{This function subsets data related to a single word when many other words are present in a dataframe}
\usage{
word_subset(reshaped_df, word_to_subset)
}
\arguments{
\item{word_to_subset}{}
}
\value{

}
\description{
This function subsets data related to a single word when many other words are present in a dataframe
}
